requirements.txt
pip==22.0.2
python-box==7.1.1
setuptools==59.6.0
